+++ Samples~/Scripts/Wwise/GetBankExample.cs
@@ -1,11 +1,15 @@
 using AK.Wwise;
 using MaJiang.Utils;
 using SaintsField.Wwise;
+using UnityEngine;
+using Event = AK.Wwise.Event;
 
 namespace SaintsField.Samples.Scripts.Wwise
 {
     public class GetBankExample : SaintsMonoBehaviour
     {
+        [SerializeField, GetWwise("PlayBGM")] private AK.Wwise.Event _playEvent;
+
         [GetWwise("BGM*")]
         public Bank bank;